+++ codemp/client/cl_parse.cpp
@@ -255,7 +255,7 @@ void CL_ParseSnapshot( msg_t *msg ) {
 	// read areamask
 	len = MSG_ReadByte( msg );
 
-	if(len > sizeof(newSnap.areamask))
+	if((unsigned)len > sizeof(newSnap.areamask))
 	{
 		Com_Error (ERR_DROP,"CL_ParseSnapshot: Invalid size %d for areamask", len);
 		return;
@@ -447,7 +447,7 @@ void CL_SystemInfoChanged( void ) {
 			gameSet = qtrue;
 		}
 
-		if((cvar_flags = Cvar_Flags(key)) == CVAR_NONEXISTENT)
+		if((unsigned)(cvar_flags = Cvar_Flags(key)) == CVAR_NONEXISTENT)
 			Cvar_Get(key, value, CVAR_SERVER_CREATED | CVAR_ROM);
 		else
 		{
@@ -731,7 +731,7 @@ void CL_ParseDownload ( msg_t *msg ) {
 	}
 
 	size = /*(unsigned short)*/MSG_ReadShort ( msg );
-	if (size < 0 || size > sizeof(data))
+	if (size < 0 || size > (int)sizeof(data))
 	{
 		Com_Error(ERR_DROP, "CL_ParseDownload: Invalid size %d for download chunk", size);
 		return;
